overhaul cspace module
This adds a large amount of missing structures for tracking the
layout of a cspace in a similar way to the kernel, ie knowing
what guards and radix bits are. Also add helpers for encoding
and decoding CPtrs and getting CPtrs into a Window. This makes
implementing an allocator relatively easy - the hard bit becomes
tracking free space, as it should be.